build: Migate to ES2016 and electron-forge
