+++ lib/src/highlighters/ace.js
@@ -1,4 +1,4 @@
-global.window = {} // Ace needs the global window object :shrug:
+if (typeof window === 'undefined') global.window = {} // Ace needs the global window object :shrug:
 const path = require('path')
 const fs = require('fs')
 const { promisify } = require('util')
