+++ src/renderer/App.js
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react'
-import './App.css'
+// import './App.css'
 import Main from './Main'
 import Preferences from './Preferences'
