+++ src/renderer/Baloon.js
@@ -2,7 +2,11 @@ import React, { PropTypes } from 'react'
 import { Label } from 'react-desktop/macOs'
 
 function Baloon({ language, relevance }) {
-  return <Label>{language}, {relevance}</Label>
+  return (
+    <Label>
+      {language}, {relevance}
+    </Label>
+  )
 }
 
 Baloon.propTypes = {
