+++ src/renderer/DynamicPreview.js
