+++ src/renderer/Main.js
