+++ src/renderer/ShortcutRecorder.js
