+++ src/renderer/StaticPreview.js
