+++ src/renderer/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Electron App</title>
+</head>
+<body>
+    <div id="root">
+        <h1>Initializing...</h1>
+    </div>
+    <script src="index.js"></script>
+</body>
+</html>
