+++ src/shared/constants/defaults.js
@@ -1,6 +1,5 @@
-'use strict'
-
-const DEFAULT_SETTINGS = {
+// eslint-disable-next-line import/prefer-default-export
+export const DEFAULT_SETTINGS = Object.freeze({
   shortcut: 'Command+Alt+V',
   fontface: 'Courier New',
   fontsize: 64, // 32pt
@@ -9,8 +8,4 @@ const DEFAULT_SETTINGS = {
   lastUsedLanguage: 'javascript',
   lineNumbers: false,
   autopaste: true
-}
-
-module.exports = {
-  DEFAULT_SETTINGS
-}
+})
